(feat) added cblas configuration
