+++ benchmarking/walltimes/fastvec_food_reviews.txt
@@ -1,5 +1,16 @@
-preprocess_reviews: 0.001080 Seconds
-preprocess_reviews: 0.001163 Seconds
-build_training_set: 1.147340 Seconds
-train_on_food_reviews: 2.490764 Seconds
-inference: 0.000144 Seconds
+preprocess_reviews: 0.001067 Seconds
+build_training_set: 0.040874 Seconds
+train_on_food_reviews: 2.480689 Seconds
+inference: 0.000192 Seconds
+preprocess_reviews: 0.001165 Seconds
+build_training_set: 0.032810 Seconds
+train_on_food_reviews: 2.281621 Seconds
+inference: 0.000171 Seconds
+preprocess_reviews: 0.001008 Seconds
+build_training_set: 0.032676 Seconds
+train_on_food_reviews: 2.255839 Seconds
+inference: 0.000143 Seconds
+preprocess_reviews: 0.001159 Seconds
+build_training_set: 0.035319 Seconds
+train_on_food_reviews: 2.012356 Seconds
+inference: 0.000150 Seconds
