+++ benchmarking/walltimes/gensim_food_reviews.txt
@@ -1,10 +1,20 @@
-preprocess_reviews: 0.024089 Seconds
-build_training_set: 0.013383 Seconds
-train_on_food_reviews: 0.084278 Seconds
-preprocess_reviews: 0.006664 Seconds
-inference: 0.003417 Seconds
-preprocess_reviews: 0.023150 Seconds
-build_training_set: 0.013266 Seconds
-train_on_food_reviews: 0.085143 Seconds
-preprocess_reviews: 0.006546 Seconds
-inference: 0.003461 Seconds
+preprocess_reviews: 0.024100 Seconds
+build_training_set: 0.013404 Seconds
+train_on_food_reviews: 0.084302 Seconds
+preprocess_reviews: 0.006783 Seconds
+inference: 0.003499 Seconds
+preprocess_reviews: 0.023869 Seconds
+build_training_set: 0.013007 Seconds
+train_on_food_reviews: 0.083756 Seconds
+preprocess_reviews: 0.006543 Seconds
+inference: 0.003405 Seconds
+preprocess_reviews: 0.022056 Seconds
+build_training_set: 0.005724 Seconds
+train_on_food_reviews: 0.046360 Seconds
+preprocess_reviews: 0.006249 Seconds
+inference: 0.003394 Seconds
+preprocess_reviews: 0.023195 Seconds
+build_training_set: 0.006163 Seconds
+train_on_food_reviews: 0.046447 Seconds
+preprocess_reviews: 0.006417 Seconds
+inference: 0.003482 Seconds
